hotfix/integration

+++ bff/web/schemas/catalog_schemas.py
@@ -1,9 +1,8 @@
 from datetime import datetime
-from decimal import Decimal
 from typing import List
 from uuid import UUID
 
-from pydantic import BaseModel, EmailStr
+from pydantic import BaseModel, EmailStr, Field
 
 
 class ProviderCreate(BaseModel):
@@ -45,7 +44,7 @@ class ProductCreate(BaseModel):
         description="Product category"
     )
     sku: str = Field(..., min_length=1, max_length=100, description="Product SKU (unique identifier)")
-    price: Decimal = Field(..., gt=0, description="Product price (must be greater than 0)")
+    price: float = Field(..., gt=0, description="Product price (must be greater than 0)")
 
 
 class ProductResponse(BaseModel):
@@ -54,7 +53,7 @@ class ProductResponse(BaseModel):
     name: str
     category: str  # Human-readable Spanish format from catalog service
     sku: str
-    price: Decimal
+    price: float
     created_at: datetime
     updated_at: datetime
